+++ platform/apps/decidim/terraform/authentik.tf
@@ -2,7 +2,7 @@ module "authentik" {
   source = "../../../../modules/authentik/oidc-app"
   count  = var.enabled ? 1 : 0
 
-  application_name = "Decidim"
+  application_name = var.application_name
   application_slug = local.slug
   category_group   = var.category_group
   icon_url         = var.icon_url